updating styles
