+++ ts/types/CreditCard.ts
@@ -52,14 +52,14 @@ export function getCardTypeFromPAN(pan: string): CreditCardType {
 
 export function getIconByCardType(type: CreditCardType) {
   const icons: { [key in CreditCardType]?: any } = {
-  "MASTERCARD": require("../../../img/wallet/cards-icons/mastercard.png"),
-  "VISA": require("../../../img/wallet/cards-icons/visa.png"),
-  "AMEX": require("../../../img/wallet/cards-icons/amex.png"),
-  "DINERS": require("../../../img/wallet/cards-icons/diners.png"),
-  "MAESTRO": require("../../../img/wallet/cards-icons/maestro.png"),
-  "VISAELECTRON": require("../../../img/wallet/cards-icons/visa-electron.png"),
-  "POSTEPAY": require("../../../img/wallet/cards-icons/postepay.png"),
-  "UNKNOWN": require("../../../img/wallet/cards-icons/unknown.png")
+  "MASTERCARD": require("../../img/wallet/cards-icons/mastercard.png"),
+  "VISA": require("../../img/wallet/cards-icons/visa.png"),
+  "AMEX": require("../../img/wallet/cards-icons/amex.png"),
+  "DINERS": require("../../img/wallet/cards-icons/diners.png"),
+  "MAESTRO": require("../../img/wallet/cards-icons/maestro.png"),
+  "VISAELECTRON": require("../../img/wallet/cards-icons/visa-electron.png"),
+  "POSTEPAY": require("../../img/wallet/cards-icons/postepay.png"),
+  "UNKNOWN": require("../../img/wallet/cards-icons/unknown.png")
   };
   return (icons[type] !== undefined) ? icons[type] : icons["UNKNOWN"];
 }
